dedicated file for KtLightAnnotations testing
